requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.6
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.0
iniconfig==1.1.1
Markdown==3.4.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
